Move sampling_conversions to transforms.
