+++ bin/inference/pycbc_inference_plot_samples
@@ -25,8 +25,8 @@ import matplotlib as mpl; mpl.use("Agg")
 import matplotlib.pyplot as plt
 import pycbc
 from pycbc import results
+from pycbc import transforms
 from pycbc.inference import option_utils
-from pycbc.inference import sampling_conversions
 import sys
 
 # command line usage
@@ -71,13 +71,13 @@ for i, arg in enumerate(parameters):
     for j in range(fp.nwalkers):
 
         # plot each walker as a different line on the subplot
-        file_parameters, cs = sampling_conversions.get_conversions(
+        file_parameters, cs = transforms.get_conversions(
                                                  parameters, fp.variable_args)
         y = fp.read_samples(file_parameters, walkers=j,
                             thin_start=opts.thin_start,
                             thin_interval=opts.thin_interval,
                             thin_end=opts.thin_end)
-        y = sampling_conversions.apply_conversions(y, cs)
+        y = transforms.apply_conversions(y, cs)
         axs[i].plot(y[arg], alpha=0.25)
 
         # y-axis label
