+++ pycbc/transforms.py
